+++ src/flask/ctx.py
@@ -471,10 +471,10 @@ class AppContext:
 
         try:
             if self._request is not None:
-                self.app.do_teardown_request(exc)
+                self.app.do_teardown_request(self, exc)
                 self._request.close()
         finally:
-            self.app.do_teardown_appcontext(exc)
+            self.app.do_teardown_appcontext(self, exc)
             _cv_app.reset(self._cv_token)
             self._cv_token = None
             appcontext_popped.send(self.app, _async_wrapper=self.app.ensure_sync)